fix append slice
